requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.1

# Graph (MVP)
networkx==3.2.1
python-louvain==0.16

# Task Queue
celery==5.3.6
APScheduler==3.10.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
python-multipart==0.0.6

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Utilities
structlog==24.1.0
python-dotenv==1.0.0
aiohttp==3.9.3
httpx==0.26.0
requests==2.31.0
pybreaker==0.7.0
bleach==6.1.0
urllib3==2.1.0

# ML & Data Science
scikit-learn==1.4.0
scipy==1.12.0
numpy==1.26.3
pandas==2.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
locust==2.20.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Monitoring
prometheus-client==0.19.0
